int/HPatel/Project_final/requirements.txt
pandas
numpy
scikit-learn
mlflow
joblib
pyyaml
